LayoutTests/platform/gtk/imported/w3c/web-platform-tests/fetch/api/basic/request-headers-worker-expected.txt


FAIL Fetch with GET assert_equals: Request has header origin: http://localhost:8800 expected (string) "http://localhost:8800" but got (object) null
FAIL Fetch with HEAD assert_equals: Request has header origin: http://localhost:8800 expected (string) "http://localhost:8800" but got (object) null
FAIL Fetch with HEAD with body promise_test: Unhandled rejection with value: object "TypeError: Type error"
PASS Fetch with PUT without body 
PASS Fetch with PUT with body 
PASS Fetch with POST without body 
PASS Fetch with POST with body 
PASS Fetch with Chicken 
PASS Fetch with Chicken with body 
